Add recurring scheduled tasks

+++ app/operations/tasks/delete_operation.rb
@@ -1,7 +1,7 @@
 module Tasks
   class DeleteOperation < BaseOperation
     def call(instance:)
-      instance.destroy!
+      instance.destroy! if instance.persisted?
     end
   end
 end
